extracted IdentifierReferenceResolver class, simplified IdentifierReferenceListener
